Refactoring. Part 1 - Move processors into separate packages
